# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Asgard",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/asgard/public/"
        },
        {
            "url": "http://wsinterno.medicamentos.gob.sv/"
        }
    ],
    "paths": {
...
        "/api/refreshauthtk": {
            "get": {
                "tags": [
                    "Login"
                ],
                "summary": "Endpoint para hacer un refresh al token.",
                "operationId": "App\\Http\\Controllers\\JWT\\AuthenticationJWTController::refreshAuthTk",
                "responses": {
                    "200": {
                        "description": "Se retorna el nuevo token."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "JWT",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}